\name{SPIA-internal}
\alias{combfunc}
\alias{getP2}
\title{Internal SPIA functions}
\description{
  Internal SPIA functions. \code{combfunc} combines two p-values into a global 
  significance p-value. \code{getP2} computes the product of two independent p-values
  that corresponds to a given global p-value. 
}
\usage{
combfunc(p1,p2)
getP2(t)
}
\details{
  These are not to be called directly by the user.
}
\keyword{internal}
