\name{Vessels}
\alias{Vessels}
\alias{DE_Vessels}
\alias{ALL_Vessels}


\title{Results from a microarray expriment comparing umbilical veins and arteries tissues}

\description{
  The \code{Vessels} dataset consists an named vector \code{DE_Vessels}
  , which represents the log2 fold changes of the genes chosen as differentially expressed between 
  ubilical veins and arteries tissue (Kim et al, 2008), and
 the universe of all Entrez gene IDs available on the array, \code{ALL_Vessels}.
 The microarray platform used was Illumina's Human-6 v2 expression BeadChip. 
 }

\usage{data(Vessels)}

\source{These data was produced at the Perinatology Research Branch, of Wayne State University (Detroit), and
accompanies the publication:\cr
 Kim JS, Romero R, Tarca A, Lajeunesse C, Han YM, Kim MJ, Suh YL, Draghici S, Mittal P, Gotsch F, Kusanovic JP, Hassan S, Kim CJ,
 Gene expression profiling demonstrates a novel role for fetal fibrocytes and the umbilical vessels in human fetoplacental development,
 J Cell Mol Med, 2008, PMID: 18298660.
}

\keyword{datasets}


