\name{SPIA-internal}
\alias{getP2}
\title{Internal SPIA functions}
\description{
  Internal SPIA functions. \code{getP2} given a pG returns two equal p-values such as  combfunc(p1,p2)=pG. 
}
\usage{
getP2(pG,combine="fisher")
}
\details{
  These are not to be called directly by the user.
}
\keyword{internal}
